Factor out output related code
