+++ sgrep_lint/semgrep/sgrep_main.py
@@ -1,6 +1,5 @@
 import argparse
 import collections
-import itertools
 import json
 import subprocess
 import sys
@@ -11,14 +10,12 @@ from pathlib import PurePath
 from typing import Any
 from typing import DefaultDict
 from typing import Dict
-from typing import Iterable
 from typing import Iterator
 from typing import List
 from typing import Optional
 from typing import Set
 from typing import Tuple
 
-import colorama
 import requests
 import semgrep.config_resolver
 import yaml
@@ -30,6 +27,8 @@ from semgrep.constants import SGREP_PATH
 from semgrep.evaluation import build_boolean_expression
 from semgrep.evaluation import enumerate_patterns_in_boolean_expression
 from semgrep.evaluation import evaluate_expression
+from semgrep.output import build_normal_output
+from semgrep.output import build_output_json
 from semgrep.sgrep_types import BooleanRuleExpression
 from semgrep.sgrep_types import InvalidRuleSchema
 from semgrep.sgrep_types import OPERATORS
@@ -50,18 +49,6 @@ from semgrep.util import print_msg
 
 SGREP_RULES_HOME = "https://github.com/returntocorp/sgrep-rules"
 MISSING_RULE_ID = "no-rule-id"
-
-
-def parse_sgrep_output(
-    sgrep_findings: List[Dict[str, Any]]
-) -> Dict[PatternId, List[SgrepRange]]:
-    output: DefaultDict[PatternId, List[SgrepRange]] = collections.defaultdict(list)
-    for finding in sgrep_findings:
-        check_id = finding["check_id"]
-        # restore the pattern id: the check_id was encoded as f"{rule_index}.{pattern_id}"
-        pattern_id = PatternId(".".join(check_id.split(".")[1:]))
-        output[pattern_id].append(sgrep_finding_to_range(finding))
-    return dict(output)
 
 
 def sgrep_finding_to_range(sgrep_finding: Dict[str, Any]) -> SgrepRange:
@@ -152,19 +139,6 @@ def invoke_sgrep(
             errors.extend(output_json["errors"])
             outputs.extend(output_json["matches"])
     return {"matches": outputs, "errors": errors}
-
-
-def fetch_lines_in_file(
-    path: Path, start_line_number: int, end_line_number: int
-) -> Optional[Iterable[str]]:
-    """
-    `line_number` is one-indexed! Returns the line if it can be found, returns None if the path doesn't exist
-    TODO: cachine
-    """
-    if not path.exists():
-        return None
-    with path.open(buffering=1) as fin:  # buffering=1 turns on line-level reads
-        return list(itertools.islice(fin, start_line_number - 1, end_line_number))
 
 
 def rewrite_message_with_metavars(
@@ -351,97 +325,6 @@ def post_output(output_url: str, output_data: Dict[str, Any]) -> None:
     print_msg(f"posting to {output_url}...")
     r = requests.post(output_url, json=output_data)
     debug_print(f"posted to {output_url} and got status_code:{r.status_code}")
-
-
-def build_output_json(output_json: Dict[str, Any]) -> str:
-    # wrap errors under "data" entry to be compatible with
-    # https://docs.r2c.dev/en/latest/api/output.html#errors
-    errors = output_json["errors"]
-    if errors:
-        output_json["errors"] = {
-            "data": output_json["errors"],
-            "message": "SgrepRuntimeErrors",
-        }
-    return json.dumps(output_json)
-
-
-def color_line(
-    line: str,
-    line_number: int,
-    start_line: int,
-    start_col: int,
-    end_line: int,
-    end_col: int,
-) -> str:
-    start_color = 0 if line_number > start_line else start_col
-    # column offset
-    start_color = max(start_color - 1, 0)
-    end_color = end_col if line_number >= end_line else len(line) + 1 + 1
-    end_color = max(end_color - 1, 0)
-    line = (
-        line[:start_color]
-        + colorama.Style.BRIGHT
-        + line[start_color : end_color + 1]  # want the color to include the end_col
-        + colorama.Style.RESET_ALL
-        + line[end_color + 1 :]
-    )
-    return line
-
-
-def finding_to_line(finding: Dict[str, Any], color_output: bool) -> Iterator[str]:
-    path = finding.get("path")
-    start_line = finding.get("start", {}).get("line")
-    end_line = finding.get("end", {}).get("line")
-    start_col = finding.get("start", {}).get("col")
-    end_col = finding.get("end", {}).get("col")
-    if path and start_line:
-        file_lines = fetch_lines_in_file(Path(path), start_line, end_line)
-        if file_lines:
-            for i, line in enumerate(file_lines):
-                if color_output:
-                    yield f"{colorama.Fore.GREEN}{start_line + i}{colorama.Style.RESET_ALL}:{color_line(line.rstrip(), start_line + i, start_line, start_col, end_line, end_col)}"
-                else:
-                    yield f"{start_line + i}:{line.rstrip()}"
-
-
-def build_normal_output(
-    output_data: Dict[str, Any], color_output: bool
-) -> Iterator[str]:
-    results = output_data.get("results", [])
-    last_file = None
-    last_message = None
-    for finding in sorted(
-        results,
-        key=lambda k: (k.get("path", "<no path>"), k.get("check_id", "<no rule id>")),
-    ):
-        RESET_COLOR = colorama.Style.RESET_ALL if color_output else ""
-        GREEN_COLOR = colorama.Fore.GREEN if color_output else ""
-        YELLOW_COLOR = colorama.Fore.YELLOW if color_output else ""
-        BLUE_COLOR = colorama.Fore.BLUE if color_output else ""
-
-        current_file = finding.get("path", "<no path>")
-        check_id = finding.get("check_id")
-        extra = finding.get("extra", {})
-        message = extra.get("message")
-        fix = extra.get("fix")
-        if last_file is None or last_file != current_file:
-            if last_file is not None:
-                yield ""
-            yield f"{GREEN_COLOR}{current_file}{RESET_COLOR}"
-            last_message = None
-        # don't display the rule line if the check is empty
-        if (
-            check_id
-            and check_id != "-"
-            and (last_message is None or last_message != message)
-        ):
-            yield f"{YELLOW_COLOR}rule:{check_id}: {finding.get('extra', {}).get('message')}{RESET_COLOR}"
-
-        last_file = current_file
-        last_message = message
-        yield from finding_to_line(finding, color_output)
-        if fix:
-            yield f"{BLUE_COLOR}autofix:{RESET_COLOR} {fix}"
 
 
 def r2c_error_format(sgrep_errors_json: Dict[str, Any]) -> Dict[str, Any]:
@@ -637,6 +520,18 @@ def main(args: argparse.Namespace) -> Dict[str, Any]:
     output_data = handle_output(outputs_after_booleans, sgrep_errors, fixes, args)
 
     return output_data
+
+
+def parse_sgrep_output(
+    sgrep_findings: List[Dict[str, Any]]
+) -> Dict[PatternId, List[SgrepRange]]:
+    output: DefaultDict[PatternId, List[SgrepRange]] = collections.defaultdict(list)
+    for finding in sgrep_findings:
+        check_id = finding["check_id"]
+        # restore the pattern id: the check_id was encoded as f"{rule_index}.{pattern_id}"
+        pattern_id = PatternId(".".join(check_id.split(".")[1:]))
+        output[pattern_id].append(sgrep_finding_to_range(finding))
+    return dict(output)
 
 
 def resolve_sgrep_output(by_rule_index: Any, all_rules: Any, args: Any) -> Any:
